+++ starknet_py/constants.py
@@ -19,4 +19,6 @@ DEVNET_FEE_CONTRACT_ADDRESS = (
 
 OZ_PROXY_STORAGE_KEY = get_storage_var_address("Proxy_implementation_hash")
 
-DEFAULT_DEPLOYER_ADDRESS = "0x041a78e741e5aF2fEc34B695679bC6891742439f7AFB8484Ecd7766661aD02BF"
+DEFAULT_DEPLOYER_ADDRESS = (
+    "0x041a78e741e5aF2fEc34B695679bC6891742439f7AFB8484Ecd7766661aD02BF"
+)
